implement `T` option (timeout) for `KeyWait`

+++ src/run/thread.cr
@@ -157,7 +157,7 @@ module Run
 			Util::AhkString.parse_key_combinations_to_charcodemap(str, @runner.settings.escape_char, @runner.x11, &block)
 		end
 
-		def parse_letter_options(str, &block : Char, Int32? -> _)
+		def parse_letter_options(str, &block : Char, Float64? -> _)
 			Util::AhkString.parse_letter_options(str, @runner.settings.escape_char, &block)
 		end
 		def parse_word_options(str)
